+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- **Run show workspace name**: Automatically include workspace relation in `run show` so WorkspaceName is always populated
+
 ## [0.5.0] - 2026-02-22
 
 ### Added